feat: add routing to demo app
