+++ projects/demo/src/app/app.component.html
@@ -1,89 +1,1 @@
-<klp-form [formGroup]="myForm" errorMessageLocation="rightOfCaption">
-	<div style="height: 500px; overflow: auto">
-		<app-deep-input>
-		</app-deep-input>
-		<div style="width: 200px">
-			<klp-form-element caption="Name Je weet erfjwefwewkfjew" spaceDistribution="34-66" direction="vertical">
-				<klp-form-text-input formControlName="name">
-				</klp-form-text-input>
-				<klp-form-error error="async"></klp-form-error>
-			</klp-form-element>
-		</div>
-		<klp-form-element caption="radioButtons" spaceDistribution="34-66">
-			<klp-form-radio formControlName="radioOption" [options]="radioOptions">
-			</klp-form-radio>
-		</klp-form-element>
-		{{myForm.get('emails').value}}
-		<klp-form-element caption="Emails" spaceDistribution="34-66">
-			<klp-form-text-input formControlName="emails" [innerValueChangeInterceptor]="innerValueChangeInterceptor" (onBlur)="blurry()">
-			</klp-form-text-input>
-		</klp-form-element>
-		<klp-form-element caption="disabledButRendered" spaceDistribution="34-66">
-			<klp-form-text-input formControlName="disabledButRendered" [innerValueChangeInterceptor]="innerValueChangeInterceptor" (onBlur)="blurry()">
-			</klp-form-text-input>
-		</klp-form-element>
-		<klp-form-element *ngIf="false" caption="unrendered" spaceDistribution="34-66">
-			<klp-form-text-input formControlName="unrendered" [innerValueChangeInterceptor]="innerValueChangeInterceptor" (onBlur)="blurry()">
-			</klp-form-text-input>
-		</klp-form-element>
-		<klp-form-element caption="Yes or no" spaceDistribution="34-66">
-			<klp-form-toggle formControlName="yesno">
-			</klp-form-toggle>
-		</klp-form-element>
-		<klp-form-element caption="Yes or no2">
-			<klp-form-toggle formControlName="yesno2">
-			</klp-form-toggle>
-		</klp-form-element>
-		<div style="width: 400px; margin-left: 500px;">
-			<klp-form-element caption="select something" style="margin-top: 200px; margin-bottom: 100px;">
-				<klp-form-select prefix="smeerpijp" [truncateOptions]="false" [options]="options" [multiple]="true" placeholder="Pick" formControlName="selector" (onSearch)="search($event)" (onEndReached)="loadMore()" (onBlur)="logSomething()" (onEnterKey)="enterKeyPressed()"></klp-form-select>
-			</klp-form-element>
-		</div>
-		<klp-form-element>
-			<klp-form-text-input formControlName="emails"></klp-form-text-input>
-		</klp-form-element>
-		<ng-container *ngFor="let email of emails; index as idx">
-			<klp-sub-form [injectInto]="subbies" [at]="idx">
-				<app-sub-form-example></app-sub-form-example>
-			</klp-sub-form>
-		</ng-container>
-		<klp-sub-form *ngIf="showSubForm" [injectInto]="myForm" at="oli">
-			blaaaa
-			<app-sub-form-example></app-sub-form-example>
-		</klp-sub-form>
-		<klp-form-submit-button [before]="beforeSubmit" [after]="afterSubmit" [submitCallback]="submitForm">
-			Submit
-		</klp-form-submit-button>
-		<klp-form-button variant="redOutlined" size="large" [disabled]="true">Red</klp-form-button>
-	</div>
-</klp-form>
-
-<klp-form-select [options]="options" [multiple]="true"></klp-form-select>
-{{testDate}}
-<klp-form-date-time-picker [(ngModel)]="testDate" [multiple]="true" [disabled]="true"></klp-form-date-time-picker>
-<klp-form-checkbox [(ngModel)]="isChecked"></klp-form-checkbox>
-<klp-form-checkbox [(ngModel)]="isChecked" [renderUndefinedAsIndeterminate]="true"></klp-form-checkbox>
-<klp-form-file-input style="height: 100%;" [(ngModel)]="myfile" [multiple]="false" [onlyShowUploadButton]="false" (ngModelChange)="fileSelected($event)" [useFullParentSize]="false"></klp-form-file-input>
-<klp-form-button (click)="toggie()">clickie</klp-form-button>
-{{groupedItems | json}}
-<klp-form-sortable-grouped-items [(ngModel)]="groupedItems"></klp-form-sortable-grouped-items>
-
-{{minutes}}
-<klp-form-hour-minute-input [(ngModel)]="minutes"></klp-form-hour-minute-input>
-<klp-form-hour-minute-input [(ngModel)]="minutes"></klp-form-hour-minute-input>
-{{fancyDate}}
-<klp-form-date-time-picker [(ngModel)]="fancyDate"></klp-form-date-time-picker>
-
-<klp-form-select [options]="staticOptions" [(ngModel)]="kers">
-	<ng-template klpSelectOptionTpl let-item="item">
-		<div style="background-color: red">custom stuff {{item | json}}</div>
-	</ng-template>
-</klp-form-select>
-
-<div style="width: 300px; position: absolute">
-	<klp-form-select [options]="staticOptions" orientation="vertical"></klp-form-select>
-</div>
-
-<klp-form-button [clickCallback]="slowClick">
-	async button
-</klp-form-button>
+<router-outlet></router-outlet>
