+++ projects/demo/src/app/app.module.ts
@@ -1,4 +1,5 @@
 import {BrowserModule} from '@angular/platform-browser';
+import { RouterModule, Routes } from '@angular/router';
 import {NgModule} from '@angular/core';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 
@@ -7,11 +8,21 @@ import {NgxEnhancyFormsModule} from '@klippa/ngx-enhancy-forms';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {SubFormExampleComponent} from './subForm/sub-form-example.component';
 import DeepInputComponent from './deep-input/deep-input.component';
+import { DemoComponent } from './demo/demo.component';
+
+const routes: Routes = [
+	{
+		path: '',
+		pathMatch: 'full',
+		component: DemoComponent,
+	}
+];
 
 @NgModule({
 	imports: [
 		BrowserModule,
 		BrowserAnimationsModule,
+		RouterModule.forRoot(routes),
 		FormsModule,
 		ReactiveFormsModule,
 		NgxEnhancyFormsModule,
@@ -20,6 +31,7 @@ import DeepInputComponent from './deep-input/deep-input.component';
 		AppComponent,
 		SubFormExampleComponent,
 		DeepInputComponent,
+		DemoComponent,
 	],
 	providers: [],
 	bootstrap: [AppComponent]
